% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkmeans.R
\name{rkmeans}
\alias{rkmeans}
\title{Title}
\usage{
rkmeans(
  data,
  assignation = "kmeans",
  update = "kmeans",
  distance = "euc",
  ranking_rule = "bordaCount"
)
}
\arguments{
\item{rankingRule}{}
}
\value{

}
\description{
Title
}
\examples{
If the update step is done using rkmeans, the new centers of the clusters
will be always within the original dataset.
Therefore, the matrix (for kmeans) or matrices (for rkmeans) of distances
between the points only need to be calculated once on the method starts.
If the update step is done using classic kmeans, then the centers will be
new objects because the will be computed as the medium points of the
objects that belong to the cluster, and thus a new matrix/matrices need
to be calculated.
}
